refactor: move and cleanup landing page related components
