+++ src/components/GenericPagesComponents/LandingPageHeroTitle/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./LandingPageHeroTitle";
