+++ src/components/GenericPagesComponents/LandingPageQuote/index.tsx
@@ -1,0 +1,1 @@
+export { LandingPageQuote } from "./LandingPageQuote";
