+++ src/components/GenericPagesComponents/LandingPageTextAndMedia/index.tsx
@@ -1,0 +1,1 @@
+export { LandingPageTextAndMedia } from "./LandingPageTextAndMedia";
