+++ src/components/GenericPagesComponents/LandingPageTextBlock/index.tsx
@@ -1,0 +1,1 @@
+export { LandingPageTextBlock } from "./LandingPageTextBlock";
