+++ src/pages/lp/[landingPageSlug].tsx
@@ -1,16 +1,15 @@
 import { GetServerSideProps, GetServerSidePropsResult, NextPage } from "next";
 
-import { getSeoProps } from "../../browser-lib/seo/getSeoProps";
-import CMSClient from "../../node-lib/cms";
-import { LandingPage } from "../../common-lib/cms-types/landingPage";
-import { LandingPageTextAndMedia } from "../../components/pages/LandingPages/LandingPageTextAndMedia";
-import { Quote } from "../../components/pages/LandingPages/Quote";
-import { SignupPrompt } from "../../components/pages/LandingPages/SignupPrompt";
-import { LandingPageTextBlock } from "../../components/pages/LandingPages/LandingPageTextBlock";
-import LandingPageHero from "../../components/pages/LandingPages/LandingPageHero";
-import getPageProps from "../../node-lib/getPageProps";
-import { getABTestedLandingPage } from "../../node-lib/cms/ab-testing";
-
+import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
+import CMSClient from "@/node-lib/cms";
+import { LandingPage } from "@/common-lib/cms-types/landingPage";
+import { LandingPageTextAndMedia } from "@/components/GenericPagesComponents/LandingPageTextAndMedia";
+import { LandingPageQuote } from "@/components/GenericPagesComponents/LandingPageQuote";
+import { SignupPrompt } from "@/components/pages/LandingPages/SignupPrompt";
+import { LandingPageTextBlock } from "@/components/GenericPagesComponents/LandingPageTextBlock";
+import LandingPageHero from "@/components/GenericPagesComponents/LandingPageHero";
+import getPageProps from "@/node-lib/getPageProps";
+import { getABTestedLandingPage } from "@/node-lib/cms/ab-testing";
 import Layout from "@/components/SharedComponents/Layout";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 
@@ -40,7 +39,7 @@ const Landing: NextPage<LandingPageProps> = ({ pageData }) => {
               }
               if (content.type == "LandingPageQuoteBlock") {
                 return (
-                  <Quote
+                  <LandingPageQuote
                     key={`${index}:${content.quote.text}`}
                     {...content.quote}
                   />
